# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/cards:
    post:
      tags:
        - Управление банковскими картами
      summary: Создание карты
      operationId: createCard
      requestBody:
        description: Запрос на создание карты
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            examples:
              Запрос на создание карты:
                description: Запрос на создание карты
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  cardNumber: '**** **** **** 3241'
                  userId: 12323213
                  validityPeriod: '2026-10-10'
                  status: ACTIVE
                  balance: 10000
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
              examples:
                Ответ с данными карты:
                  description: Ответ с данными карты
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    cardInfoList:
                      - cardId: 123233241
                        cardNumber: '**** **** **** 3241'
                        ownerUsername: Nick
                        ownerFirstName: Joe
                        ownerLastName: Doe
                        validityPeriod: '2025-10-10'
                        status: ACTIVE
                        balance: 10000
                Ответ об успешном удалении или запросе на блокировку:
                  description: Ответ об успешном удалении или запросе на блокировку
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Карта не найдена
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
    delete:
      tags:
        - Управление банковскими картами
      summary: Удаление карты
      operationId: deleteCard
      requestBody:
        description: Запрос на удаление карты
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            examples:
              Запрос на удаление карты:
                description: Запрос на удаление карты
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  cardNumber: '**** **** **** 3241'
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
              examples:
                Ответ с данными карты:
                  description: Ответ с данными карты
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    cardInfoList:
                      - cardId: 123233241
                        cardNumber: '**** **** **** 3241'
                        ownerUsername: Nick
                        ownerFirstName: Joe
                        ownerLastName: Doe
                        validityPeriod: '2025-10-10'
                        status: ACTIVE
                        balance: 10000
                Ответ об успешном удалении или запросе на блокировку:
                  description: Ответ об успешном удалении или запросе на блокировку
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Карта не найдена
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
    patch:
      tags:
        - Управление банковскими картами
      summary: Изменение данных карты
      operationId: updateCard
      requestBody:
        description: Запрос на изменение данных карты
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            examples:
              Запрос на изменение данных карты:
                description: Запрос на изменение данных карты
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  cardNumber: '**** **** **** 3241'
                  status: ACTIVE
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
              examples:
                Ответ с данными карты:
                  description: Ответ с данными карты
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    cardInfoList:
                      - cardId: 123233241
                        cardNumber: '**** **** **** 3241'
                        ownerUsername: Nick
                        ownerFirstName: Joe
                        ownerLastName: Doe
                        validityPeriod: '2025-10-10'
                        status: ACTIVE
                        balance: 10000
                Ответ об успешном удалении или запросе на блокировку:
                  description: Ответ об успешном удалении или запросе на блокировку
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Карта не найдена
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
  /api/v1/cards/search:
    post:
      tags:
        - Управление банковскими картами
      summary: Поиск банковских карт пользователей
      operationId: searchCard
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: direction
          in: query
          required: true
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - name: sortField
          in: query
          required: true
          schema:
            type: string
      requestBody:
        description: Поиск банковских карт с фильтрацией и пагинацией
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            examples:
              Запрос на поиск банковских карт:
                description: Запрос на поиск банковских карт
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  cardId: 1232312
                  cardNumber: '**** **** **** 3241'
                  ownerUsername: SuperMan
                  ownerFullName: Joe Doe
                  status: ACTIVE
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
              examples:
                Ответ с данными карты:
                  description: Ответ с данными карты
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    cardInfoList:
                      - cardId: 123233241
                        cardNumber: '**** **** **** 3241'
                        ownerUsername: Nick
                        ownerFirstName: Joe
                        ownerLastName: Doe
                        validityPeriod: '2025-10-10'
                        status: ACTIVE
                        balance: 10000
                Ответ об успешном удалении или запросе на блокировку:
                  description: Ответ об успешном удалении или запросе на блокировку
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Карта не найдена
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
  /api/v1/cards/block-request:
    post:
      tags:
        - Управление банковскими картами
      summary: Запрос пользователя на блокировку карты
      operationId: blockRequestCard
      requestBody:
        description: Запрос пользователя на блокировку карты
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
            examples:
              Запрос пользователя на блокировку карты:
                description: Запрос пользователя на блокировку карты
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  cardNumber: '**** **** **** 3241'
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
              examples:
                Ответ с данными карты:
                  description: Ответ с данными карты
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    cardInfoList:
                      - cardId: 123233241
                        cardNumber: '**** **** **** 3241'
                        ownerUsername: Nick
                        ownerFirstName: Joe
                        ownerLastName: Doe
                        validityPeriod: '2025-10-10'
                        status: ACTIVE
                        balance: 10000
                Ответ об успешном удалении или запросе на блокировку:
                  description: Ответ об успешном удалении или запросе на блокировку
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 91a0a5a8-9428-46b2-961e-d14a8485b48b
                    errorMessage: 'Ошибка авторизации: Access Denied'
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Карта не найдена
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        errorMessage:
          type: string
...
    CardRequest:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        cardId:
          type: integer
          format: int64
        cardNumber:
          type: string
        ownerUsername:
          type: string
        ownerFullName:
          type: string
        userId:
          type: integer
          format: int64
        validityPeriod:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
    CardInfo:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        cardNumber:
          type: string
        ownerUsername:
          type: string
        ownerFirstName:
          type: string
        ownerLastName:
          type: string
        validityPeriod:
          type: string
          format: date
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
    CardResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        cardInfoList:
          type: array
          items:
            $ref: '#/components/schemas/CardInfo'
...
  securitySchemes:
    JWT:
      type: http
      in: header
      scheme: bearer
      bearerFormat: 'Bearer '